refactor(lints): Consolidate unknown_lints
