+++ src/cargo/core/workspace.rs
@@ -36,10 +36,10 @@ use crate::lints::rules::redundant_homepage;
 use crate::lints::rules::redundant_readme;
 use crate::lints::rules::text_direction_codepoint_in_comment;
 use crate::lints::rules::text_direction_codepoint_in_literal;
+use crate::lints::rules::unknown_lints;
 use crate::lints::rules::unused_build_dependencies_no_build_rs;
 use crate::lints::rules::unused_workspace_dependencies;
 use crate::lints::rules::unused_workspace_package_fields;
-use crate::lints::unknown_lints;
 use crate::ops;
 use crate::ops::lockfile::LOCKFILE_NAME;
 use crate::sources::{CRATES_IO_INDEX, CRATES_IO_REGISTRY, PathSource, SourceConfigMap};
